Fixed crash in DatabaseHandler when using getProfile, deleteProfile with invalid city

+++ src/com/example/PrayerTimes/DatabaseHandler.java
@@ -82,8 +82,8 @@ public class DatabaseHandler extends SQLiteOpenHelper {
 		Cursor cursor = db.query(TABLE_CITIES, new String[] { KEY_ID,
 				KEY_NAME, KEY_LAT, KEY_LONG, KEY_TZ, KEY_useGPS, KEY_useTimezone }, KEY_NAME + "=?",
 				new String[] { cityName }, null, null, null, null);
-		if (cursor != null){
-			cursor.moveToFirst();
+		if (cursor != null && cursor.moveToFirst()){
+
 			profile.cityName = cursor.getString(1);
 			profile.savedLatitude = Double.parseDouble(cursor.getString(2));
 			profile.savedLongitude = Double.parseDouble(cursor.getString(3));
@@ -103,7 +103,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
 		if (cursor.moveToFirst()) {
 			do {
 				Profile profile = new Profile();
-				
+
 				profile.cityName = cursor.getString(1);
 				profile.savedLatitude = Double.parseDouble(cursor.getString(2));
 				profile.savedLongitude = Double.parseDouble(cursor.getString(3));
@@ -117,11 +117,18 @@ public class DatabaseHandler extends SQLiteOpenHelper {
 		}
 		return citiesList;
 	}
-	
+
 	public void deleteProfile(Profile profile) {
 		SQLiteDatabase db = this.getWritableDatabase();
-		db.delete(TABLE_CITIES, KEY_NAME + " = ?",
-				new String[] { profile.cityName });
+		// Check if the Profile is actually there
+		Cursor cursor = db.query(TABLE_CITIES, new String[] { KEY_ID,
+				KEY_NAME, KEY_LAT, KEY_LONG, KEY_TZ, KEY_useGPS, KEY_useTimezone }, KEY_NAME + "=?",
+				new String[] { profile.cityName }, null, null, null, null);
+		// If yes, then delete
+		if (cursor.moveToFirst()){
+			db.delete(TABLE_CITIES, KEY_NAME + " = ?",
+					new String[] { profile.cityName });
+		}
 		db.close();
 	}
 }
